infrastructure/requirements.txt
aws-cdk-lib==2.218.0
constructs>=10.0.79
python-dotenv==1.0.0
aws-cdk.aws-bedrock-agentcore-alpha